requirements.txt
azure-functions==1.21.3
dnspython==2.7.0
pymongo==4.11.2
